[REFACTOR] Use isQuest from expo.core instead of local implementation
-Transitioned expo-horizon-core to use isQuest from expo-core internally
-Transitioned expo-horizon-location to use isQuest from expo-core internallyransitioned to isQuest function provided by expo-cor
