--- expo-horizon-location/android/src/main/java/expo/modules/location/VRUtilities.kt
@@ -1,13 +1,0 @@
-package expo.modules.location
-
-import android.os.Build
-
-// This class is a local copy of `expo.modules.core.utilities.VRUtilities` to eliminate the dependency on Expo SDK 54.
-// TODO: In the future, move to use `expo.modules.core.utilities.VRUtilities` when possible.
-class VRUtilities {
-  companion object {
-    fun isHorizon(): Boolean {
-      return (Build.MANUFACTURER.equals("Oculus", ignoreCase = true) || Build.MANUFACTURER.equals("Meta", ignoreCase = true)) && Build.MODEL.contains("Quest")
-    }
-  }
-}
